feat: Added board configuration file and refactored squares.
